repair: Introduce scylla repair script
The idea behind the script is to split "big" repair into "small"
repairs.

Examples:
1) Run repair on node 127.0.0.{1,2,3} for both primary range and
non-primary range
./scylla_repair.py --keyspace ks3 --hosts
127.0.0.1,127.0.0.2,127.0.0.3 --ports 7199,7200,7300 --pr --npr

2) Run with --cont option to rerun repair for the failed range
./scylla_repair.py --keyspace ks3 --hosts
127.0.0.1,127.0.0.2,127.0.0.3 --ports 7199,7200,7300 --cont

3) Print the repair cmd for primary range for node 127.0.0.3
./scylla_repair.py --keyspace ks3 --hosts 127.0.0.3 --ports 7300
--dryrun --pr

The script is capable of running repair for other nodes in the cluster.
However, this requires the scylla JMX port (7199) to be listened on
non-localhost.

For now, we can recommend user to run the scrpit on all the nodes
instead running the script on a single node to repair the whole cluster.
For example to repair keyspace 'myks' on node 1, node 2, node3.

On node1:
./scylla_repair.py --keyspace myks --pr --hosts broadcast_address_of_node1
On node2:
./scylla_repair.py --keyspace myks --pr --hosts broadcast_address_of_node2
On node3:
./scylla_repair.py --keyspace myks --pr --hosts broadcast_address_of_node3

Message-Id: <61d85a879fea4d465695ba3f0ee483f3dc77f7cb.1495491725.git.asias@scylladb.com>