scylla-enterprise: Add encrypted file impl
Very simple block-encrypting file wrapper.

Uses user provided symmetric key + ESSIV block IV calculation
to encrypt data.

All encryption is done unpadded. To handle file sizes we use
a slightly shaky scheme:

Since all writes are assumed to be done by us, and must be aligned,
we can assume in turn that any resizing should be made by our truncate
method. If we attept to truncate to a size not a multiple of our
_key_ block size (typically 16), we add the same size to the actual
truncation size.
On read we then check the file size. If we're reading from a file
with unaliged size, we know there are key-block-size junk at the end.
We can align down the last decryption call to match block size, then
discard the excessive bytes from the result.

If we're in a read/write situation, we need to keep size updated, and
we could possibly race with disk op/continuations.
But we are really only for ro/wo cases.