scylla-enterprise: commitlog/hints encryption.
Adds extension to encrypt commitlog/hints files.

Since commitlog segments have no metadata (persistent or otherwise),
the extension adds .-files to the commitlog dir, one for each file
opened for write, into which the encryption parameters are written.

On replay, the same .-file is opened, and the properties are
used to find the appropriate key for decryption.