cql3: Add overload for query processor
The many overloads for functions in `cql3::query_processor` can be
somewhat confusing.

Four variables are the kind of `client_state` used to invoke the
`execute` function of the CQL statement (`internal` or `external`), the
choice of which of `execute` or `execute_internal` gets invoked, whether
`check_access` of the statements gets called, and the result type.

A particular overload, `cql3::query_processor::process(_,
db::consistency_level, std::initializer_list)`, has the following
properties:
- It uses `internal` `client_state`.
- `execute` gets called.
- `check_access` is not invoked.
- The result type is `cql3::untyped_result_set`.

This overload is used pervasively in `auth/` as a means to execute CQL
statements "internally" as part of the implementation of CQL statements
for access-control. For example, executing LIST USERS as a user
internally executes a SELECT query on the `system_auth.users` table.
Since these queries are not protected with access-control, we break the
dependency loop.

In the next change, we fix a bug with the implementation of LIST USERS
which is caused because it used an overload of
`query_processor::process` that invokes `execute_internal`. What we want
for that change is most of the existing behavior of
`cql3::query_processor::process(_, db::consistency_level,
std::initializer_list)` except that we want a different return type.

Instead of copying and pasting, this change refactors that common parts
into a new function and implements the other functions in terms of it.

We also add some documentation to clarify some of the confusion around
the present (unchanged here) state of things.