Add an interface to query total memory
We want to add a component that will reserve memory on boot. Since
this memory will no longer be available to the rest of the system,
the rest of the system should not be aware of its existence. Introduce
an interface to query total memory that takes reserved memory into
account.