In memory sstable storage
This patch introduce a capability to mirror sstables in memory. When
sstable is written it is written both into a permanent storage and into
memory, read goes to memory only. During boot all sstables are brought
into memory. This is beneficial for workloads that can entirely fit
into RAM.

To enable in memory store there is a new parameter in-memory-storage-size-mb
(0 by default).

Current implementation stores only Index and Data files in memory since
rest of sstable files are either read only once or held in memory
already.