Redirect seastar submodule to upstream Seastar
Currently, we don't have enterprise-specific patches to seastar
(a property we should preserve). Since enterprise tracks master,
it should receive any seastar fixes as updates to upstream, not
as backports to a private seastar branch. We can therefore point
the seastar submodule to upstream, and not require periodic
synchronization of the enterprise-specific seastar submodule.

When enterprise is branched, we'd need to point the seastar
submodule at scylla-enterprise-seastar.git in order to be able
to backport fixes, but that occurs an a much lower frequency.