tests: add incremental compaction test
this test will check that incremental compaction releases fragments
composing a run as expected.

it provides its own lambda -- that is called by compaction every time
it replaces an input fragment with a new one -- to check that
everything is in order. For example, fragment A that precedes
fragment B in a given run must be released first if and only if
all of its data is sealead in a new fragment that will compose the
output run.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>