sstables: add incremental compaction strategy
This new compaction strategy is like size tiered compaction strategy
in nature, meaning that it also has size tiers, but the difference
is that each size tier is based on sstable runs instead of plain
individual sstables.
It also means that compaction strategy will look for N similar-sized
sstable runs to compact, not just sstables.

This new strategy uses the previously introduced incremental backlog
tracker, which accounts for runs instead and works essentially as
size tiered backlog tracker.

One thing that can improved for the future is the compaction
interface for selecting candidates. Right now, it's sstable oriented
so we have to convert candidate sstables to runs. sstable_set interface
helps us with that, but it's not very optimal.

Here it has a static fragment size of 1GB, but subsequent work will
make it a parameter which will give us more control and help testing.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>