sstables: add sstable/fragment size as compaction strategy parameter
Make it a parameter to aid with testing and also provide us a way
to change it dynamically if we want to reduce the overhead for
compaction free space.

Fragment/sstable size is set by user in mb, similar to max sstable size
for leveled compaction strategy.

Set it as follow:
ALTER TABLE users
  WITH compaction =
  {'class' : 'IncrementalCompactionStrategy', 'sstable_size_in_mb' : 1000 }

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>