service/qos: change registration of distributed keyspace to prevent a race
The fix is done by moving the registration to after the tokens are initialized.
The service_level_controller periodicaly queries an accesor for service
levels configurations (that are stored in a table). The accessor was
registered right after the system dystributed keyspace was initilialized,
but before the tokens were fully initialized (later in the function). If the
service level controller queries the service level accessor before the
token initialization, it will result in an error that indicates that
the tokens are still not initialized.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>
Message-Id: <20190305045904.16305-3-eliransin@scylladb.com>