replicated_key_provider: Force flush on key generation
Refs #754

If system info, i.e. commitlog, is using encryption via
the replicated provider, we cannot have said keys reside
only in commitlog.

To deal with this we force-flush every key generated to
sstable. This means creating a write key is much more expensive,
but since we should in most cases hit the cache, the end result
should be acceptable.