retire scyllarepair
As the scyllamanager gains ground, we should retire scyllarepair.
We still use scyllarepair sometimes, but as the manager grows in
capability we do this less and less.

In this patch I propose we remove it from enterprise master. We should
announce its deprecation for 2019.1. This should be time enough.

This script was extremely helpful early on. It goes away with great
respect.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20190318225803.7243-1-glauber@scylladb.com>