audit: mark user as anonymous if it has no name in audit log
Recent changes to the authenticated user returns an empty optional
for the user name if and only if this user is anonymous. This is
in contrast to the old implementation which returned "anonymous" for
the user name if the user is anonymous. The audit code didn't take
this change into accout, resulting in logging the user as "undefined"
instead. This commit fixes audit to take this under account.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>
Message-Id: <20190319120204.6546-1-eliransin@scylladb.com>