sstables: make resharding efficient for incremental compaction strategy
Default implementation was used which reshards individual sstables.
That will not work for ICS which has a high number of fixed-size
sstables as that can lead to a higher number of sstables after the
resharding is done. What we'll do to solve that is to reshard
together all fragments that belong to the same sstable run.

Fixes #761.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20190424025602.16611-1-raphaelsc@scylladb.com>