storage_service: add cluster feature for in-memory tables
In enterprise-2018.1, we added a scylla_tables.in_memory column unconditionally
(that is, without adding a gossip feature to coordinate addition of the
column). That makes upgrading to a version that doesn't support in-memory tables
impossible without breaking schema synchronization.

In enterprise-2019.1, we wish to support upgrading from both scylla-3.0 and
enterprise-2018.1. That makes it hard to choose how to send schemas: if we
send them without the in-memory column, 2018.1 nodes will fail to read the
schema, while if we send them with the in_memory column, 3.0 nodes will fail
to read the schema.

To solve this problem, we add a gossip feature and also repurpose the
lack of the xxhash feature as an indication that we're upgrading from
enterprise-2018.1. Since the lack of a feature can also indicate we haven't
gossiped yet, we also check that range_tombsones (a very old feature)
_is_ present.

We use the following logic:

 - in_memory feature present: this is a 2019.1 cluster and we can send in_memory
   columns.
 - range_tombstones not present: we don't know anything yet, so we don't send
   in_memory.
 - xxhash not present, range_tombstones present: this is a mixed 2018.1/2019.1
   cluster and we can send in_memory columns.
 - in_memory not present and xxhash present: this is a mixed 2018.1/3.0 cluster
   and we cannot send in_memory columns.

The logic fails if trying to upgrade from 2.1 (which doesn't have xxhash) - it
will be misrecognized as 2018.1. However, that release is long out of date. Users
should upgrade to 3.0 first.

(cherry picked from commit e5c1be28fb9ac80cc9c39f265d123b84b847584c)