schema: add schema_feature for in-memory tables
In-memory tables added a column to system_schema.scylla_tables, so if we
do a rolling upgrade from a version that doesn't support in memory tables to
a version that does, we must remove the in_memory column so that the old node
isn't confused.

Add a schema_feature for this, and translate schema mutations to the
old scylla_tables layout if the feature is not present.

(cherry picked from commit db732e2de0be72f5885096a50851686dc068d89b)