auth: standard_role_manager: allow disabling the creation of the roles_attribute_table
Schema upgrade tests that check for upgrating from master do not take into
consideration the existence of enterprise-only tables, so we'll have to disable
those tables for these tests.

To do that, add a member in db::config. It is easily accessible in tests and is
propagated everywhere, even if it's sad to add another #include for it.