system_distributed_keyspace: allow disabling the creation of the service_levels table
Since schema upgrade tests check against a schema that doesn't contain
enterprise-only tables, we need a way to disable creation of those tables for these
tests. This patch allows disabling the service_levels table.

Again we add a member to db::config, since it is propagated everywhere and
accessible from the tests.