sstables/ICS: Use non-enforcement of min threshold for strategy to move towards zero backlog
Non enforcement of minimum threshold is needed to help the situation where there are many
sstables inducing a large backlog, but they are not yet meeting the minimum criteria for
compaction. When that happens, the backlog never goes down. That's already done with STCS,
so we only need to do the same with ICS.

Fixes #875.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20190524051608.7960-1-raphaelsc@scylladb.com>